Add logic for context-sensitive labels
These are specified in the settings.js file as:

addCommand(1234, "setlabel", {X : ["X1", "X2"],
                              Y : ["Y1", "Y2"]})

which means that, for nodes with POS X (defined as the node label with
any dash tags chopped off), cycle through X1 and X2; for Y cycle through
Y1 and Y2.  Be careful not to let one of the choices for X not have POS
label X, or you will be booted out of that rotation.  (This could be a
lot less brittle).

The motivation for this is the case of the “a personal” in Spanish.  It
is desirable to use the same key to cycle between NP-SBJ/OB1/OB2 for
NPs, and between PP/PP-OB1/2 for PPs.