js, settings: improve toggling of dash tags
Now there is but one function to do the toggling.  There is also one
key-bindable function.  It selects a tagset based on whether the node to
be toggled is:
- a leaf node
- a clause
- otherwise
The toggleVerbExtension function is retained as an alias, for
compatibility.  Specifying extensions with dash included in settings.js
is now deprecated, but will still work.