treedrawing.contextMenu.js: don’t include extra tags in context menu
Before, the code would search 2 levels deep in the context menu groups.
So, if groups [A, B] and [B, C] are defined, rightclicking on a
node labeled A would give choices B and C.  The latter is included
because A incldes B as a choice, and there is also a group that includes
B and C.  The inclusion of C here is counterintuitive; we now no longer
do this.

Addresses part of #19